style: fix C4244/C4267 in NetworkManager/Profile/NotesLoaderBMS (item 2)
Continuing the item-2 sweep. All casts document existing intentional
behavior, no logic change -- characterization-verified.

- NetworkManager.cpp (6 sites): lua_tointeger() (lua_Integer) into int
  config fields x4; two uint64_t response sizes into lua_pushnumber's
  lua_Number x2.
- Profile.cpp (10 sites, once the newly-surfaced ones from a proper
  clean rebuild are included -- see below): GetAge()'s float constant
  (DEFAULT_BIRTH_YEAR) subtraction returned as int; count_if's
  iterator-diff returned as float; a chain of double-literal-driven
  calorie-burn math (Duration/60.0, weight/2.205, GetAge() -> float,
  final /4.184 * Duration) into float locals/return; four
  size()/Right(size_t)/lua_createtable(size_t) sites into int params.
- NotesLoaderBMS.cpp (10 sites, §5-adjacent NotesLoader* per AGENTS.md
  #5 -- characterization only, verified against the existing
  test_NotesLoaderBMS.cpp pinned values, unchanged: 26 assertions / 2
  cases still pass with identical numbers): a double measureSize into
  a float adjustedMeasureSize x2; two long long time-signature values
  into SetTimeSignatureAtRow's int params; a double beat sum into
  BeatToNoteRow's float param; an unsigned int BPM into a float
  currentBPM; four size()/find()-returned size_t values into
  int/unsigned locals or lua_c API int parameters.

Also corrected the running site count: the last two commits' "187 -> 158"
was itself measured with an incomplete (touched-files-only, not
--clean-first) rebuild and undercounted. A genuine clean rebuild with
/wd4244/4267 removed now shows the true current total: 140 unique sites
across 69 files remain (all ten files fixed so far confirmed at zero).

Verified: sm_tests 5966/226 (BMS case explicitly re-checked: 26
assertions / 2 cases, all pinned values identical), ctest 100%, Release
build + --SelfTest green, with /wd4244/4267 removed AND restored.

+++ src/NotesLoaderBMS.cpp
@@ -567,7 +567,7 @@ bool BMSChart::Load( const RString &chartPath )
 				channel -= 40;
 				flag = true;
 			}
-			int count = data.size() / 2;
+			int count = static_cast<int>( data.size() / 2 );
 			for (int i = 0; i < count; i++)
 			{
 				RString value = data.substr(2 * i, 2);
@@ -677,7 +677,7 @@ int BMSSong::AllocateKeysound( RString filename, RString path )
 		return mapKeysoundToIndex[normalizedFilename];
 	}
 
-	unsigned index = out->m_vsKeysoundFile.size();
+	unsigned index = static_cast<unsigned>( out->m_vsKeysoundFile.size() );
 	out->m_vsKeysoundFile.push_back( normalizedFilename );
 	mapKeysoundToIndex[filename] = index;
 	mapKeysoundToIndex[normalizedFilename] = index;
@@ -924,7 +924,7 @@ void BMSChartReader::ReadHeaders()
 
 void BMSChartReader::CalculateStepsType()
 {
-	nonEmptyTracksCount = nonEmptyTracks.size();
+	nonEmptyTracksCount = static_cast<int>( nonEmptyTracks.size() );
 	out->m_StepsType = DetermineStepsType();
 	if(out->m_StepsType == StepsType_Invalid)
 	{
@@ -1304,8 +1304,8 @@ bool BMSChartReader::ReadNoteData()
 			measureSize = 4.0f;
 			BMSMeasures::iterator it = in->measures.find(trackMeasure);
 			if( it != in->measures.end() ) measureSize = it->second.size * 4.0;
-			adjustedMeasureSize = measureSize;
-			if( trackMeasure < firstNoteMeasure ) adjustedMeasureSize = measureSize = 4.0f;
+			adjustedMeasureSize = static_cast<float>( measureSize );
+			if( trackMeasure < firstNoteMeasure ) adjustedMeasureSize = static_cast<float>( measureSize = 4.0f );
 
 			// measure size adjustment
 			{
@@ -1317,10 +1317,10 @@ bool BMSChartReader::ReadNoteData()
 					num /= 2;
 					den /= 2;
 				}
-				td.SetTimeSignatureAtRow( BeatToNoteRow(measureStartBeat), num, den );
+				td.SetTimeSignatureAtRow( BeatToNoteRow(measureStartBeat), static_cast<int>(num), static_cast<int>(den) );
 
 				// Since BMS measure events only last through the measure, we need to restore the default measure length.
-				td.SetTimeSignatureAtRow(BeatToNoteRow(measureStartBeat + measureSize), 4, 4);
+				td.SetTimeSignatureAtRow(BeatToNoteRow(static_cast<float>(measureStartBeat + measureSize)), 4, 4);
 			}
 			// end measure size adjustment
 		}
@@ -1334,7 +1334,7 @@ bool BMSChartReader::ReadNoteData()
 			unsigned int bpm;
 			if( sscanf(obj.value, "%x", &bpm) == 1 )
 			{
-				if( bpm > 0 ) td.SetBPMAtRow( row, measureAdjust * (currentBPM = bpm) );
+				if( bpm > 0 ) td.SetBPMAtRow( row, measureAdjust * (currentBPM = static_cast<float>(bpm)) );
 			}
 		}
 		else if( channel == 4 ) // bga change
@@ -1596,7 +1596,7 @@ void BMSSongLoader::AddToSong()
 				// XXX: This matches (double), but I haven't seen it used. Again, MORE EXAMPLES NEEDED
 				if( tag.find('l') != tag.npos )
 				{
-					unsigned pos = tag.find('l');
+					unsigned pos = static_cast<unsigned>( tag.find('l') );
 					if( pos > 2 && tag.substr(pos - 2, 4) == "solo" )
 					{
 						// (solo) -- an edit, apparently (Thanks Glenn!)
